+++ validation/graylog_rest_api.py
@@ -198,7 +198,25 @@ class GraylogRestApi:
             ],
             'matching_type': 'AND'
         }
-        return self._create_alert_rule(title, rule, 'AND', time, additional_threshold_type='<', additional_threshold=additional_threshold, second_stream=second_stream)
+        return self._create_alert_rule(title, rule, 'AND', time, additional_threshold_type='>', second_stream=second_stream)
+
+    def create_alert_rule_or(self, title, time):
+        rule = {
+            'field': 'a',
+            'type': 1,
+            'value': 'b'
+        }
+        second_stream = {
+            'field_rule': [
+                {
+                    'field': 'b',
+                    'type': 1,
+                    'value': 'titi'
+                }
+            ],
+            'matching_type': 'AND'
+        }
+        return self._create_alert_rule(title, rule, 'OR', time, additional_threshold_type='<', second_stream=second_stream)
 
     def get_alert_rule(self, name):
         response = self._get(f'plugins/com.airbus_cyber_security.graylog.wizard/alerts/{name}')